Fixed id and temporarily disabled backface culling
Fixed the id so it actually shows a different colour depending on the voxel type, and disabled backface culling for now due to not updating correctly.

+++ BuildScape/src/Chunk.h
@@ -36,7 +36,7 @@ public:
 	struct InstanceData {
 		glm::vec3 offset;
 		glm::mat4 rotation;
-		int id;
+		float id;
 	};
 	std::vector<InstanceData> instances;
